+++ src/web/services/auth/oauth2_providers/discord.rs
@@ -1,24 +1,25 @@
 //! Discord OAuth2 flow.
 
+use std::sync::Arc;
+
+use actix_web::http::header::ACCEPT;
+use chrono::{DateTime, Duration, Utc};
+use futures::future::LocalBoxFuture;
+use oauth2::{AccessToken, RefreshToken, Scope, TokenResponse};
+use oauth2::{AuthUrl, TokenUrl};
+use oauth2::basic::{BasicClient, BasicTokenResponse};
+use serenity::model::user::CurrentUser;
 use crate::api::rcos::send_query;
 use crate::api::rcos::users::accounts::reverse_lookup::ReverseLookup;
 use crate::api::rcos::users::UserAccountType;
 use crate::env::global_config;
 use crate::error::TelescopeError;
 use crate::web::services::auth::identity::{AuthenticationCookie, RootIdentity};
-use crate::web::services::auth::oauth2_providers::{Oauth2Identity, Oauth2IdentityProvider};
 use crate::web::services::auth::IdentityProvider;
-use actix_web::http::header::ACCEPT;
-use chrono::{DateTime, Duration, Utc};
-use futures::future::LocalBoxFuture;
-use oauth2::basic::{BasicClient, BasicTokenResponse};
-use oauth2::{AccessToken, RefreshToken, Scope, TokenResponse};
-use oauth2::{AuthUrl, TokenUrl};
-use serenity::model::user::CurrentUser;
-use std::sync::Arc;
+use crate::web::services::auth::oauth2_providers::{Oauth2Identity, Oauth2IdentityProvider};
 
 /// The Discord API endpoint to query for user data.
-const DISCORD_API_ENDPOINT: &'static str = "https://discord.com/api/v8";
+pub const DISCORD_API_ENDPOINT: &'static str = "https://discord.com/api/v8";
 
 /// Zero-sized type used to represent Discord based identity verification.
 pub struct DiscordOAuth;